% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optweight-package.R
\docType{package}
\name{optweight-package}
\alias{optweight-package}
\title{optweight: Optimization-Based Stable Balancing Weights}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Use optimization to estimate weights that balance covariates for binary, multi-category, continuous, and multivariate treatments in the spirit of Zubizarreta (2015) \doi{10.1080/01621459.2015.1023805}. The degree of balance can be specified for each covariate. In addition, sampling weights can be estimated that allow a sample to generalize to a population specified with given target moments of covariates, as in matching-adjusted indirect comparison (MAIC).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ngreifer.github.io/optweight/}
  \item \url{https://github.com/ngreifer/optweight}
  \item Report bugs at \url{https://github.com/ngreifer/optweight/issues}
}

}
\author{
\strong{Maintainer}: Noah Greifer \email{noah.greifer@gmail.com} (\href{https://orcid.org/0000-0003-3067-7154}{ORCID})

}
\keyword{internal}
